requirements.txt
numpy
Pillow
pyvips
shapely
ultralytics
torch
torchvision
